// src/components/contact-form-large-zsd/contact-form-large-zsd-interface.d.ts
import { LocaleVariants } from '../../interfaces/global-variants';
export interface ContactFormLargeZsdProps {
    onSubmit: (data: unknown) => void;
    headline: string;
    tagline?: string;
    locale: LocaleVariants;
    loading?: boolean;
    privacyPolicyUrl: string;
    success?: boolean;
}
//# sourceMappingURL=contact-form-large-zsd-interface.d.ts.map